+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Features:
 
 * ``use_annotations`` supports more types in the ``typing`` module, as
   well as ``dict`` and ``list``.
+* ``use_annotations`` supports `typing.Optional`.
 
 1.0.0 (2018-01-04)
 ------------------